Made sure wsapi unit tests compile.
