+++ wsapi/ack_test.go
@@ -1,0 +1,21 @@
+package wsapi_test
+
+import (
+	/*"encoding/json"
+	  "fmt"
+	  "github.com/FactomProject/factomd/common/entryBlock"
+	  "github.com/FactomProject/factomd/common/interfaces"
+	  "github.com/FactomProject/factomd/common/primitives"
+	  "github.com/FactomProject/factomd/receipts"
+	  "github.com/FactomProject/factomd/testHelper"
+	  . "github.com/FactomProject/factomd/wsapi"
+	  "github.com/FactomProject/web"
+	  "net/http"
+	  "strings"*/
+
+	"testing"
+)
+
+func TestHandleV2FactoidACK(t *testing.T) {
+
+}
